fix: 修复时间戳类型冲突和异步任务准备
主要修改：
1. 统一时间工具函数使用u64作为内部存储类型
2. 修复ai_task.rs中的current_timestamp_secs导入
3. 修复session.rs中的timestamp_to_rfc3339类型转换
4. 修复stream.rs中的now_timestamp_millis类型不匹配
5. 优化时间工具函数，提供明确的类型转换函数

这些修改为异步任务的稳定运行提供了基础，解决了时间戳类型冲突问题。

+++ alou-edge/src/router/session.rs
@@ -213,7 +213,7 @@ pub(crate) async fn handle_agent_chat(
                         error: format!(
                             "每日请求限额已用完。您今天已经使用了所有可用请求。\n\nDaily request limit exceeded. You have used all available requests for today.\n\n剩余请求 / Remaining requests: {}\n重置时间 / Reset time: {} (UTC)",
                             remaining,
-                            crate::utils::time::timestamp_to_rfc3339(next_day)
+                            crate::utils::time::timestamp_to_rfc3339(next_day as u64)
                         ),
                         remaining_requests: remaining,
                         limit_exceeded: true,
